% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_canopy.R
\name{classify_canopy}
\alias{classify_canopy}
\title{Classify Canopy Cover}
\usage{
classify_canopy(ndvi_raster, threshold = 0.4)
}
\arguments{
\item{ndvi_raster}{A NDVI raster layer.}

\item{threshold}{NDVI threshold to classify canopy.}
}
\value{
A binary raster of canopy presence (1 = canopy, 0 = no canopy).
}
\description{
Classify Canopy Cover
}
