% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_canopy_change.R
\name{detect_canopy_change}
\alias{detect_canopy_change}
\title{Detect Changes in Canopy Cover}
\usage{
detect_canopy_change(canopy_early, canopy_late)
}
\arguments{
\item{canopy_early}{Raster of early canopy cover.}

\item{canopy_late}{Raster of later canopy cover.}
}
\value{
A raster showing gain/loss/no-change in canopy.
}
\description{
Detect Changes in Canopy Cover
}
