% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_satellite_data.R
\name{download_satellite_data}
\alias{download_satellite_data}
\title{Download and Load Satellite Images}
\usage{
download_satellite_data(file_path_2018, file_path_2023)
}
\arguments{
\item{file_path_2018}{Path to 2018 image file.}

\item{file_path_2023}{Path to 2023 image file.}
}
\value{
A list containing two SpatRaster objects: img2018 and img2023.
}
\description{
Download and Load Satellite Images
}
