% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenology_metrics.R
\name{extract_phenology_metrics}
\alias{extract_phenology_metrics}
\title{Extract Phenology Metrics}
\usage{
extract_phenology_metrics(ndvi_ts, dates, method = "derivative")
}
\arguments{
\item{ndvi_ts}{A time-series of NDVI values (vector).}

\item{dates}{A corresponding vector of dates.}

\item{method}{Smoothing method (default: derivative).}
}
\value{
List containing onset DOY and the smoothed NDVI curve.
}
\description{
Extract Phenology Metrics
}
