% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate Automated Report}
\usage{
generate_report(aoi_name)
}
\arguments{
\item{aoi_name}{Name for the Area of Interest (AOI) report.}
}
\value{
Renders a HTML report to the working directory.
}
\description{
Generate Automated Report
}
