% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_change_map.R
\name{leaflet_change_map}
\alias{leaflet_change_map}
\title{Create Leaflet Map of Canopy Change}
\usage{
leaflet_change_map(change_raster)
}
\arguments{
\item{change_raster}{Raster showing canopy change.}
}
\value{
A Leaflet interactive map object.
}
\description{
Create Leaflet Map of Canopy Change
}
