% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_map.R
\name{phenology_map}
\alias{phenology_map}
\title{Create Map of Autumn Onset}
\usage{
phenology_map(ndvi_stack, date_vector)
}
\arguments{
\item{ndvi_stack}{A SpatRaster stack of NDVI images.}

\item{date_vector}{Vector of dates matching each layer.}
}
\value{
Raster of estimated autumn onset day of year (DOY).
}
\description{
Create Map of Autumn Onset
}
