% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenology_curve.R
\name{plot_phenology_curve}
\alias{plot_phenology_curve}
\title{Plot Phenology Curve}
\usage{
plot_phenology_curve(curve_data, onset_doy)
}
\arguments{
\item{curve_data}{A dataframe containing NDVI curve and DOY.}

\item{onset_doy}{Estimated onset day of year (autumn).}
}
\value{
A ggplot2 plot.
}
\description{
Plot Phenology Curve
}
